onProgres : state management

+++ lib/src/features/devices/data/light.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:smarthome_algeria/src/features/devices/domain/device.dart';
+import 'package:smarthome_algeria/src/features/devices/data/devices.dart';
 
 
 enum LightPowerConsumption { low, medium, high }
 
 extension LightPowerConsumptionExtension on LightPowerConsumption {
-  double get powerConsumption {
+  int get powerConsumption {
     switch (this) {
       case LightPowerConsumption.low:
         return 10;
@@ -18,7 +18,7 @@ extension LightPowerConsumptionExtension on LightPowerConsumption {
 }
 
 class Light extends Device {
-  Light({required double powerConsumption, String name = 'Light', Color color = Colors.yellow, required int id})
+  Light({required int powerConsumption, String name = 'Light', Color color = Colors.yellow, required int id})
       : super(
             type: DeviceType.light,
             name: name,
